Reverse the process of splitting bwd kernel into multiple files.
This makes the review easier.
